Revert "Fixed NumberOfKids() return type"
Those changes are correct but no longer necessary in this branch
because they were specific to the list-based PageStack implementation.

This reverts branch commit 43d92c4a7850e6f4fc9c1cfcc07cf4a4c1d7fc7f.

+++ src/ipc/Kids.cc
@@ -49,7 +49,7 @@ void Kids::init()
         storage.push_back(Kid(kid_name));
     }
 
-    Must(storage.size() == NumberOfKids());
+    Must(storage.size() == static_cast<size_t>(NumberOfKids()));
 }
 
 /// returns kid by pid
